extract `async-loader` package
